+++ R/betaSandwich-package.R
@@ -1,0 +1,7 @@
+#' @aliases betaSandwich-package
+#' @keywords internal
+"_PACKAGE"
+
+## usethis namespace: start
+## usethis namespace: end
+NULL