+++ NEWS.md
@@ -1,10 +1,29 @@
-# betaDelta 1.0.0.9000
+# betaSandwich 1.0.3.9000
 
 Latest development version.
 
-* Added the `DiffBetaDelta()` function.
+* Added the `DiffBetaSandwich()` and the `RSqBetaSandwich()` functions.
 
-# betaDelta 1.0.0
+# betaSandwich 1.0.3
+
+## Patch
+
+* Added the `BetaADF()` function.
+
+# betaSandwich 1.0.2
+
+## Patch
+
+* Factor variables are dummy coded.
+
+# betaSandwich 1.0.1
+
+## Patch
+
+* Added the `nas1982` data set.
+* Minor edits to methods.
+
+# betaSandwich 1.0.0
 
 ## Major